conf: Report error if invalid type specified for character device
If invalid type is specified, e.g.
<serial type='foo'>
    <target port='0'/>
</serial>

We replace 'foo' with "null" type implicitly, without reporting an
error message to tell the user, and "start" or "edit" the domain
will be success.

It's not good to guess what the user wants, This patch is to fix
the problem.

* src/conf/domain_conf.c
(cherry picked from commit 92d653010606a45a69ea6edd55ba689bffca0b44)

Signed-off-by: Steve Hodgson <shodgson@solarflare.com>