report error when specifying wrong desturi
When we do peer2peer migration, the dest uri is an address of the
target host as seen from the source machine. So we must specify
the ip or hostname of target host in dest uri. If we do not specify
it, report an error to the user.

Signed-off-by: Wen Congyang <wency@cn.fujitsu.com>
(cherry picked from commit 59d13aae329ce7d4153e5f8a7d7ec94b779a610b)

Signed-off-by: Steve Hodgson <shodgson@solarflare.com>