vf-hotplug: Move vlan from an element to an attribute.
Whilst vlan is really a property of the interface rather than
the VF, libvirt appears to have been moving in the direction
that vlans are part of the network configuration rather than
the VM configuration. We need vlans for the VF only, so it
makes sense to have them as a VF attribute.

This is also slightly "safer" w.r.t upstream libvirt changes,
since the elements is not in the interface namespace.

Signed-off-by: Steve Hodgson <shodgson@solarflare.com>
(cherry picked from commit 7082a757ec5d1b4b470cd578ae51284443ece125)

Conflicts:

	src/conf/domain_conf.c
	src/qemu/qemu_command.c
	src/qemu/qemu_driver.c
	src/qemu/qemu_migration.c
	src/qemu/qemu_process.c